separated aboutview from focusview
