+++ frontend/app/src/config/ModeDispatcher.tsx
@@ -7,7 +7,7 @@ export enum AppMode {
 }
 
 const hosts: { [k: string]: AppMode } = {
-  'localhost:3000': AppMode.TRIBES,
+  'localhost:3005': AppMode.COMMUNITY,
   'localhost:13000': AppMode.TRIBES,
   'localhost:23000': AppMode.TRIBES,
   'tribes.sphinx.chat': AppMode.TRIBES,
